Add more tests and comments.

+++ Tests/Realm.Tests/Sync/SynchronizedInstanceTests.cs
@@ -358,21 +358,24 @@ namespace Realms.Tests.Sync
         {
             SyncTestHelpers.RunBaasTestAsync(async () =>
             {
-                var partition = Guid.NewGuid().ToString();
-
-                var originalConfig = await GetIntegrationConfigAsync(partition);
+                var originalPartition = Guid.NewGuid().ToString();
+                var originalConfig = await GetIntegrationConfigAsync(originalPartition);
                 if (originalEncrypted)
                 {
                     originalConfig.EncryptionKey = TestHelpers.GetEncryptionKey(42);
                 }
 
-                var copyConfig = await GetIntegrationConfigAsync(partition);
+                var copiedPartition = Guid.NewGuid().ToString();
+                var copyConfig = await GetIntegrationConfigAsync(copiedPartition);
                 if (copyEncrypted)
                 {
                     copyConfig.EncryptionKey = TestHelpers.GetEncryptionKey(14);
                 }
 
-                File.Delete(copyConfig.DatabasePath);
+                if (!allowOverwrite)
+                {
+                    File.Delete(copyConfig.DatabasePath);
+                }
 
                 using var originalRealm = GetRealm(originalConfig);
 
@@ -382,9 +385,47 @@ namespace Realms.Tests.Sync
                 await WaitForDownloadAsync(originalRealm);
 
                 originalRealm.WriteCopyWithoutClientFileId(copyConfig, allowOverwrite);
+
+                using var copiedlRealm = GetRealm(copyConfig);
+
+                Assert.AreEqual(copiedlRealm.All<ObjectIdPrimaryKeyWithValueObject>().Count(), DummyDataSize / 2);
             });
         }
 
+        [Test]
+        public void WriteCopyWithoutFileId_FailsWhenNotFinished([Values(true, false)] bool originalEncrypted,
+                                                                 [Values(true, false)] bool copyEncrypted,
+                                                                 [Values(true, false)] bool allowOverwrite)
+        {
+            SyncTestHelpers.RunBaasTestAsync(async () =>
+            {
+                var originalPartition = Guid.NewGuid().ToString();
+                var originalConfig = await GetIntegrationConfigAsync(originalPartition);
+                if (originalEncrypted)
+                {
+                    originalConfig.EncryptionKey = TestHelpers.GetEncryptionKey(42);
+                }
+
+                var copiedPartition = Guid.NewGuid().ToString();
+                var copyConfig = await GetIntegrationConfigAsync(copiedPartition);
+                if (copyEncrypted)
+                {
+                    copyConfig.EncryptionKey = TestHelpers.GetEncryptionKey(14);
+                }
+
+                if (!allowOverwrite)
+                {
+                    File.Delete(copyConfig.DatabasePath);
+                }
+
+                using var originalRealm = GetRealm(originalConfig);
+
+                AddDummyData(originalRealm, true);
+
+                // The error is thrown as a generic `RealmError` by Core which translates to a generic `RealmException` on our side.
+                Assert.Throws<RealmException>(() => originalRealm.WriteCopyWithoutClientFileId(copyConfig, allowOverwrite));
+            });
+        }
 
         private const int DummyDataSize = 100;
